core, eth, les: move bloomIndexer to core package
